+++ Renderer/resource/BillBoard.h
@@ -14,7 +14,7 @@ public:
 	// 빌보드 여러개 생성 
 	void MakeBillBoard(const std::vector<BillBoardVertex>& vertices);
 	// 움직이지 않는 빌보드 생성 ( NonSpritable ) 
-	void MakeBillBoard(const DirectX::XMFLOAT3& position,UINT halfWidth, UINT height, TextureIndex image);
+	void MakeBillBoard(const DirectX::XMFLOAT3& position,float halfWidth, float height, TextureIndex image);
 
 	void Render(ComPtr<ID3D12GraphicsCommandList>& commandList, D3D12_GPU_DESCRIPTOR_HANDLE texHeap, D3D12_GPU_VIRTUAL_ADDRESS cameraBuffer);
 private:
